test(agent): move the patched-iterator case to a dispositioned sibling
The patched-array-iterator regression landed in src/agent/factory.test.ts,
a hermetic unit file. Its `Object.defineProperty(Array.prototype, ...)`
makes the file effect-bearing, so `lint:test-semantic-dispositions`
failed with "missing semantic disposition", and its
`createEphemeralAgent({ id, tools })` call omitted the required `system`,
so `lint:test-typecheck` reported a new TS2769 for that entry point.

Move the case verbatim into
tests/integration/agent/ag-ui-runtime-restrictions-intrinsics.test.ts —
the dispositioned sibling that already patches intrinsics to prove this
PR's tool ceiling — and give the rebuilt config the `system` it needs.
The test still fails against the previous `for...of` loop over the
captured entries, so it keeps covering the fix.
